requirements.txt
anyio==3.6.2
appdirs==1.4.4
async-generator==1.10
attrs==22.2.0
certifi
cffi==1.15.1
charset-normalizer==2.1.1
consonance==0.1.5
cryptography==39.0.0
dissononce==0.34.3
et-xmlfile==1.1.0
exceptiongroup==1.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
numpy==1.21.6
openai
openpyxl==3.0.10
outcome==1.2.0
pandas==1.3.5
pandas-stubs==1.2.0.62
protobuf==4.21.12
PyAutoIt==0.6.5
pycparser==2.21
PySocks==1.7.1
python-axolotl==0.2.2
python-axolotl-curve25519==0.4.1.post2
python-dateutil==2.8.2
python-dotenv
python-telegram-bot==20.0
pytz==2022.7
readline==6.2.4.1
requests==2.28.1
rfc3986==1.5.0
selenium==4.7.2
six==1.10.0
sniffio==1.3.0
sortedcontainers==2.4.0
tqdm==4.64.1
transitions==0.9.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
urllib3==1.26.13
wsproto==1.2.0
yowsup==3.3.0
